LayoutTests/fast/dom/intersection-observer-document-leak-expected.txt

main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
Tests that IntersectionObserver doesn't cause Document leaks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isDocumentAlive(frameDocumentID) is true
PASS internals.isDocumentAlive(frameDocumentID) is true
PASS internals.isDocumentAlive(frameDocumentID) is true
PASS At least 1 iframe document was destroyed.
PASS successfullyParsed is true

TEST COMPLETE
